Final tweaks

+++ EbanxExam.Domain/Models/Account.cs
@@ -3,6 +3,6 @@
   public class Account
   {
     public string Id { get; set; }
-    public decimal? Balance { get; set; }
+    public int? Balance { get; set; }
   }
 }
